Removed Rhai references
